requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.1.3
django-environ==0.11.2
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
l18n==2021.3
laces==0.1.1
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pillow_heif==0.21.0
psycopg2-binary==2.9.10
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
telepath==0.3.1
urllib3==2.2.3
wagtail==6.3.1
wheel==0.44.0
Willow==1.9.0
